+++ .env_SAMPLE
@@ -47,9 +47,11 @@ export ALLOW_ADMIN_URL=True
 #export AKAMAI_CLIENT_TOKEN=<akamai_client_token>
 #export AKAMAI_FAST_PURGE_URL=<akamai_fast_purge_url>
 #export AKAMAI_PURGE_ALL_URL=<akamai_purge_all_url>
+#export AKAMAI_PURGE_HOSTNAMES=<akamai_purge_hostnames>
 
 # export ENABLE_CLOUDFRONT_CACHE_PURGE=True
 # export CLOUDFRONT_DISTRIBUTION_ID_FILES=<cloudfront_distribution_id_files_cf_gov>
+#export CLOUDFRONT_PURGE_HOSTNAMES=<cloudfront_purge_hostnames>
 
 #########################################################################
 # Postgres Environment Vars.
