+++ cfgov/cdntools/backends.py
@@ -2,30 +2,13 @@ import json
 import logging
 import os
 
-from django.conf import settings
-from django.contrib.auth.models import User
-from django.db import models
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-
 from wagtail.contrib.frontend_cache.backends import BaseBackend
-from wagtail.contrib.frontend_cache.utils import PurgeBatch
-from wagtail.documents.models import Document
 
 import requests
 from akamai.edgegrid import EdgeGridAuth
 
-from v1.models.images import CFGOVRendition
-
 
 logger = logging.getLogger(__name__)
-
-
-class CDNHistory(models.Model):
-    created = models.DateTimeField(auto_now_add=True)
-    subject = models.CharField(max_length=2083)
-    message = models.CharField(max_length=255)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
 
 
 class AkamaiBackend(BaseBackend):
@@ -132,19 +115,13 @@ class AkamaiDeletingBackend(AkamaiBackend):
         )
 
 
-@receiver(post_save, sender=Document)
-@receiver(post_save, sender=CFGOVRendition)
-def cloudfront_cache_invalidation(sender, instance, **kwargs):
-    if not settings.ENABLE_CLOUDFRONT_CACHE_PURGE:
-        return
+class MockCacheBackend(BaseBackend):
+    def __init__(self, params):
+        super().__init__(params)
+        self.cached_urls = params.get("CACHED_URLS")
 
-    if not instance.file:
-        return
+    def purge(self, url):
+        self.cached_urls.append(url)
 
-    url = instance.file.url
-
-    logger.info(f'Purging {url} from "files" cache')
-
-    batch = PurgeBatch()
-    batch.add_url(url)
-    batch.purge(backends=["files"])
+    def purge_all(self):
+        self.cached_urls.append("__all__")
